# src/main/java/gov/tn/dhs/quarkus/box/model/UploadFileResponse.java
package gov.tn.dhs.quarkus.box.model;

import com.fasterxml.jackson.annotation.JsonInclude;
import com.fasterxml.jackson.annotation.JsonProperty;

@JsonInclude(JsonInclude.Include.NON_NULL)
public class UploadFileResponse {

  @JsonProperty("status")
  private String status;

  @JsonProperty("fileId")
  private String fileId;

  public String getStatus() {
    return status;
  }

  public void setStatus(String status) {
    this.status = status;
  }

  public String getFileId() {
    return fileId;
  }

  public void setFileId(String fileId) {
    this.fileId = fileId;
  }

}