# src/main/resources/box-api-Swagger20.yaml
swagger: '2.0'
info:
  version: '1.0.0'
  title: box_api
  description: Box API for users
  contact: {}
host: www.example.com
basePath: /
schemes:
- https
consumes:
- application/json
produces:
- application/json
paths:
  /folder:
    get:
      description: Get file
      summary: Search for documents in a folder
      tags:
      - Folder
      operationId: Searchfordocumentsinafolder
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: appUserId
        in: query
        required: true
        type: string
        description: app user id
      - name: folderId
        in: query
        required: true
        type: string
        description: folder id
      - name: fileName
        in: query
        required: false
        type: string
        description: folder id
      - name: offset
        in: query
        required: false
        type: integer
        format: int32
        description: offset
      - name: limit
        in: query
        required: false
        type: integer
        format: int32
        description: limit
      responses:
        '200':
          description: File or Folder obtained
          schema:
            $ref: '#/definitions/SearchResult'
          headers: {}
        '400':
          description: App User or File/Folder not found
          schema:
            $ref: '#/definitions/ServiceError'
          headers: {}
        '404':
          description: File/Folder not found
          schema:
            $ref: '#/definitions/ServiceError'
          headers: {}
        '500':
          description: Service error
          schema:
            $ref: '#/definitions/ServiceError'
          headers: {}
    post:
      description: Create folder
      summary: Create folder
      tags:
      - Folder
      operationId: Createfolder
      deprecated: false
      produces:
      - application/json
      consumes:
      - application/json
      parameters:
      - name: body
        in: body
        required: true
        description: ''
        schema:
          $ref: '#/definitions/FolderCreationRequest'
      responses:
        '200':
          description: Successfully created new folder
          schema:
            $ref: '#/definitions/FolderCreationResponse'
          headers: {}
        '400':
          description: Some of the parameters are missing or not valid.
          schema:
            $ref: '#/definitions/ServiceError'
          headers: {}
        '404':
          description: The parent folder could not be found, or the authenticated user does not have access to the parent folder.
          schema:
            $ref: '#/definitions/ServiceError'
          headers: {}
        '409':
          description: The folder is locked or the folder already exists.
          schema:
            $ref: '#/definitions/ServiceError'
          headers: {}
  /file:
    get:
      description: Get file
      summary: Download file
      tags:
      - File
      operationId: Downloadfile
      deprecated: false
      produces:
      - application/octet-stream
      - application/json
      parameters:
      - name: appUserId
        in: query
        required: true
        type: string
        description: app user id
      - name: fileId
        in: query
        required: true
        type: string
        description: file id
      responses:
        '200':
          description: A binary file
          schema:
            type: string
            format: binary
          headers: {}
        '202':
          description: File is not ready to be downloaded.
          schema:
            type: string
          headers:
            Retry-After:
              description: Time in seconds after which the file will be available for the client to download.
              type: string
        '404':
          description: File not found
          schema:
            type: string
          headers: {}
    post:
      description: Upload file
      summary: Upload file
      tags:
      - File
      operationId: Uploadfile
      deprecated: false
      produces:
      - application/json
      consumes:
      - multipart/form-data
      parameters:
      - name: appUserId
        in: query
        required: true
        type: string
        description: app user id
      - name: boxFolderId
        in: query
        required: true
        type: string
        description: box folder id
      - name: file
        in: formData
        required: true
        type: file
        format: file
        description: ''
      responses:
        '200':
          description: Document ID of uploaded content is returned
          schema:
            $ref: '#/definitions/UploadFileResponse'
          headers: {}
        '400':
          description: Folder error
          schema:
            $ref: '#/definitions/ServiceError'
          headers: {}
        '409':
          description: File exists error
          schema:
            $ref: '#/definitions/ServiceError'
          headers: {}
    delete:
      description: Delete file
      summary: Delete file
      tags:
      - File
      operationId: Deletefile
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: appUserId
        in: query
        required: true
        type: string
        description: app user id
      - name: fileId
        in: query
        required: true
        type: string
        description: file id
      responses:
        '200':
          description: File deleted successfully.
          schema:
            $ref: '#/definitions/DocumentDeletionResult'
          headers: {}
        '400':
          description: App User or File/Folder not found
          schema:
            $ref: '#/definitions/ServiceError'
          headers: {}
        '404':
          description: File/Folder not found.
          schema:
            $ref: '#/definitions/ServiceError'
          headers: {}
        '500':
          description: You cannot delete this file as it is already attached to a Case or is Verified.
          schema:
            $ref: '#/definitions/ServiceError'
          headers: {}
  /file-view:
    get:
      description: View file
      summary: View file
      tags:
      - File
      operationId: Viewfile
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: appUserId
        in: query
        required: true
        type: string
        description: app user id
      - name: fileId
        in: query
        required: true
        type: string
        description: file id
      responses:
        '200':
          description: view metadata
          schema:
            $ref: '#/definitions/DocumentViewResponse'
          headers: {}
        '400':
          description: Some of the parameters are missing or not valid
          schema:
            $ref: '#/definitions/ServiceError'
          headers: {}
        '404':
          description: File not found
          schema:
            $ref: '#/definitions/ServiceError'
          headers: {}
        '500':
          description: document view error
          schema:
            $ref: '#/definitions/ServiceError'
          headers: {}
  /file-link:
    get:
      description: Link File
      summary: Link File
      tags:
      - File
      operationId: LinkFile
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: appUserId
        in: query
        required: true
        type: string
        description: app user id
      - name: fileId
        in: query
        required: true
        type: string
        description: file id
      responses:
        '200':
          description: view metadata
          schema:
            $ref: '#/definitions/DocumentLinkResponse'
          headers: {}
        '400':
          description: Some of the parameters are missing or not valid
          schema:
            $ref: '#/definitions/ServiceError'
          headers: {}
        '404':
          description: File not found
          schema:
            $ref: '#/definitions/ServiceError'
          headers: {}
        '500':
          description: document link error
          schema:
            $ref: '#/definitions/ServiceError'
          headers: {}
definitions:
  SearchResult:
    title: SearchResult
    type: object
    properties:
      complete:
        $ref: '#/definitions/Complete'
      fileData:
        type: array
        items:
          $ref: '#/definitions/FileInfo'
  FolderCreationRequest:
    title: FolderCreationRequest
    type: object
    properties:
      firstName:
        description: First name
        type: string
      lastName:
        description: Last name
        type: string
      mpiid:
        type: string
      logonuserid:
        type: string
    required:
    - firstName
    - lastName
    - mpiid
  FolderCreationResponse:
    title: FolderCreationResponse
    type: object
    properties:
      appUserId:
        description: The unique identifier for the user the folder is associated with.
        type: string
      folderId:
        description: The unique identifier that represents the folder.
        type: string
  UploadFileResponse:
    title: UploadFileResponse
    type: object
    properties:
      status:
        description: Status message for operation
        type: string
      fileId:
        description: File ID of uploaded content
        type: string
  ServiceError:
    title: ServiceError
    type: object
    properties:
      code:
        description: Error code
        type: string
      message:
        description: Error description
        type: string
  FileUploadData:
    title: FileUploadData
    type: object
    properties:
      file:
        type: string
        format: binary
    required:
    - file
  FileDownloadData:
    title: FileDownloadData
    type: object
    properties:
      file:
        type: string
        format: binary
  FileInfo:
    title: FileInfo
    type: object
    properties:
      fileId:
        type: string
      fileName:
        type: string
      itemType:
        $ref: '#/definitions/ItemType'
  DocumentDeletionResult:
    title: DocumentDeletionResult
    type: object
    properties:
      code:
        type: string
      message:
        type: string
  DocumentViewResponse:
    title: DocumentViewResponse
    type: object
    properties:
      previewUrl:
        type: string
      shortPreviewUrl:
        type: string
  DocumentLinkResponse:
    title: DocumentLinkResponse
    type: object
    properties:
      linkUrl:
        type: string
  Complete:
    title: Complete
    type: string
    enum:
    - 'True'
    - 'False'
  ItemType:
    title: ItemType
    type: string
    enum:
    - file
    - folder
tags:
- name: Folder
  description: ''
- name: File
  description: ''